requirements.txt
Django==1.10.2
django-crispy-forms==1.6.1
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
Markdown==2.6.7
whitenoise==3.2.3
